demux: wav: change data_pos sign
Use the same type than vlc_stream_Tell().

+++ modules/demux/wav.c
@@ -47,7 +47,7 @@ typedef struct
     es_format_t     fmt;
     es_out_id_t     *p_es;
 
-    int64_t         i_data_pos;
+    uint64_t        i_data_pos;
     int64_t         i_data_size;
 
     unsigned int    i_frame_size;
@@ -599,7 +599,8 @@ static int Open( vlc_object_t * p_this )
     if( !b_is_rf64 || i_size < UINT32_MAX )
     {
         int64_t i_stream_size = stream_Size( p_demux->s );
-        if( i_stream_size > 0 && i_stream_size >= i_size + p_sys->i_data_pos )
+        if( i_stream_size > 0
+         && (uint64_t) i_stream_size >= i_size + p_sys->i_data_pos )
             p_sys->i_data_size = i_size;
     }
